Use a direct mapped cache to reduce the lookup overhead

+++ projects/common/deception/include/Deception.h
@@ -245,7 +245,7 @@ namespace Deception {
             static constexpr auto LineMask = NumLines - 1;
             using Line_t = L;
             static constexpr uint8_t computeIndex(Address input) noexcept {
-                return (input >> Line_t::ShiftAmount) & LineMask;
+                return (static_cast<uint16_t>(input) >> Line_t::ShiftAmount) & LineMask;
             }
             void clear() noexcept {
                 for (auto line : lines) {
